Moviendo domain.model.dto a infrastructure.adapter.dto - Modificando nombres de argumentos donde se utilizaban los dtos - Formateando el codigo
