+++ src/main/java/com/entrevistador/generadorfeedback/infrastructure/adapter/dto/FeedbackComentarioDto.java
@@ -1,4 +1,4 @@
-package com.entrevistador.generadorfeedback.domain.model.dto;
+package com.entrevistador.generadorfeedback.infrastructure.adapter.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
