+++ src/main/java/com/entrevistador/generadorfeedback/infrastructure/adapter/dto/IdEntrevista.java
@@ -1,4 +1,4 @@
-package com.entrevistador.generadorfeedback.domain.model.dto;
+package com.entrevistador.generadorfeedback.infrastructure.adapter.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
