+++ src/main/java/com/entrevistador/generadorfeedback/infrastructure/rest/controller/FeedbackController.java
@@ -1,7 +1,7 @@
 package com.entrevistador.generadorfeedback.infrastructure.rest.controller;
 
 import com.entrevistador.generadorfeedback.application.usescases.FeedbackCreation;
-import com.entrevistador.generadorfeedback.domain.model.dto.FeedbackResponseDto;
+import com.entrevistador.generadorfeedback.infrastructure.adapter.dto.FeedbackResponseDto;
 import com.entrevistador.generadorfeedback.infrastructure.adapter.mapper.FeedbackMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,8 +18,7 @@ public class FeedbackController {
     private final FeedbackMapper feedbackMapper;
 
     @GetMapping(value = "/entrevistas/{entrevistaId}")
-    public Flux<FeedbackResponseDto> obtenerFeedback(
-            @PathVariable String entrevistaId) {
+    public Flux<FeedbackResponseDto> obtenerFeedback(@PathVariable String entrevistaId) {
         return this.feedbackCreation.obtenerFeedback(entrevistaId)
                 .map(this.feedbackMapper::mapFeedbackResponseToFeedbackResponseDto);
     }
