+++ src/main/java/com/entrevistador/generadorfeedback/infrastructure/rest/controller/PreguntaController.java
@@ -1,7 +1,7 @@
 package com.entrevistador.generadorfeedback.infrastructure.rest.controller;
 
 import com.entrevistador.generadorfeedback.application.usescases.PreguntaCreation;
-import com.entrevistador.generadorfeedback.domain.model.dto.PreguntaComentarioDto;
+import com.entrevistador.generadorfeedback.infrastructure.adapter.dto.PreguntaComentarioDto;
 import com.entrevistador.generadorfeedback.infrastructure.adapter.mapper.FeedbackMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,8 +19,7 @@ public class PreguntaController {
 
 
     @GetMapping(value = "/entrevistas/{entrevistaId}")
-    public Flux<PreguntaComentarioDto> obtenerPreguntas(
-            @PathVariable String entrevistaId) {
+    public Flux<PreguntaComentarioDto> obtenerPreguntas(@PathVariable String entrevistaId) {
         return this.preguntaCreation.obtenerPreguntas(entrevistaId)
                 .map(this.feedbackMapper::mapPreguntaComentarioEntrevistaToPreguntaComentarioDto);
     }
